+++ NEWS.md
@@ -63,6 +63,8 @@ mcp v0.4 is a major breaking change with the aim of remaining relatively stable 
 
 -   Default `plot(fit)` style has been updated in many ways to accommodate multiple regression and group-effects.
 
+-   Added `mcp(..., series = "data_column")` to set series in ARMA models.
+
 -   Added option to test hypotheses on the prior using `hypothesis(fit, prior = TRUE)`.
 
 -   `mcp_example(...)` now shows an illustrative plot of the fitted example by default. Disable using `plot = FALSE`.